# Use Sequence instead of []XY for DCEL edges