--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use `Sequence` instead of `[]XY` for DCEL edges.
+
 ## v0.40.1
 
 2022-11-08
@@ -7,8 +11,6 @@
 __Special thanks to @missinglink for contributing to this release.__
 
 - Add benchmarks for WKB parsing.
-
-- Use `Sequence` instead of `[]XY` for DCEL edges.
 
 ## v0.40.0