## Description

commit 08c83784f8e7d2a5ae37a7204cc96dc55a755525
Author: Peter Stace <peterstace@gmail.com>
Date:   Sun Nov 6 20:36:23 2022 +1100

```
Use Sequence instead of []XY for DCEL edges

The DCEL algorithm needs to represent the points along an edge. The
sources of these points are `Sequence` values from the input geometries.
The DCEL geometry extraction phase (the last step of the DCEL algorithm)
includes a step that inserts points along the edge into the output
geometry.

Before this change, DCEL edges represent points as a slice of `XY`. This
has the disadvantage of a conversion step from a `Sequence` to `[]XY`
(in the DCEL input stage), and then from `[]XY` back to `Sequence` (in
the DCEL output stage).

This change simplifies things by just using `Sequence` throughout. It
also slightly increases the range of points stored on each DCEL edge to
include the origin of the edge, and the final point on the next edge.
These points are redundant (since the output stage could read them from
the vertex records) but help with debugging. Because the Sequences are
shallow copies rather than deep copies, there is no performance penalty.
```

## Check List

Have you:

- Added unit tests? N/A, relies on existing.

- Add cmprefimpl tests? (if appropriate?) N/A, relies on existing.

- Updated release notes? (if appropriate?) Yes.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/302

## Benchmark Results

- Some _very_ slight perf improvements. This PR is more about code simplification than getting a perf improvement, so the perf check was more to ensure there wasn't an unexpected regression.

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:arm64
LineEnvelope/0-2                                              2.23ns ± 7%    2.31ns ± 7%   +3.57%  (p=0.014 n=14+15)
LineEnvelope/1-2                                              2.15ns ± 2%    2.13ns ± 2%     ~     (p=0.059 n=14+13)
LineEnvelope/2-2                                              2.14ns ± 3%    2.13ns ± 2%     ~     (p=0.168 n=15+14)
LineEnvelope/3-2                                              2.13ns ± 3%    2.12ns ± 2%     ~     (p=0.659 n=15+14)
MarshalWKB/polygon/n=10-2                                     92.0ns ± 2%    92.3ns ± 1%     ~     (p=0.191 n=13+12)
MarshalWKB/polygon/n=100-2                                     237ns ± 5%     237ns ± 5%     ~     (p=0.830 n=14+14)
MarshalWKB/polygon/n=1000-2                                   1.54µs ±11%    1.59µs ±11%     ~     (p=0.210 n=14+14)
MarshalWKB/polygon/n=10000-2                                  33.6µs ±44%    39.7µs ±43%     ~     (p=0.148 n=15+15)
UnmarshalWKB/polygon/n=10-2                                    159ns ± 3%     156ns ± 1%   -2.21%  (p=0.000 n=14+13)
UnmarshalWKB/polygon/n=100-2                                   300ns ± 4%     295ns ± 6%     ~     (p=0.095 n=15+14)
UnmarshalWKB/polygon/n=1000-2                                 1.80µs ±16%    1.77µs ±11%     ~     (p=0.621 n=15+14)
UnmarshalWKB/polygon/n=10000-2                                38.9µs ±47%   41.8µs ±113%     ~     (p=0.874 n=14+14)
IntersectsLineStringWithLineString/n=10-2                      767ns ± 2%     767ns ± 5%     ~     (p=0.448 n=14+14)
IntersectsLineStringWithLineString/n=100-2                    11.9µs ± 4%    11.8µs ± 3%     ~     (p=0.471 n=13+13)
IntersectsLineStringWithLineString/n=1000-2                    141µs ±10%     146µs ±21%     ~     (p=0.354 n=14+15)
IntersectsLineStringWithLineString/n=10000-2                  3.03ms ±49%    3.21ms ±32%     ~     (p=0.252 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20-2                      553ns ± 0%     555ns ± 1%   +0.44%  (p=0.001 n=13+14)
IntersectsMultiPointWithMultiPoint/n=200-2                    7.55µs ± 1%    7.53µs ± 1%     ~     (p=0.838 n=15+14)
IntersectsMultiPointWithMultiPoint/n=2000-2                   71.4µs ± 2%    71.3µs ± 1%     ~     (p=0.946 n=14+14)
IntersectsMultiPointWithMultiPoint/n=20000-2                   779µs ± 1%     783µs ± 2%   +0.54%  (p=0.022 n=13+14)
PolygonSingleRingValidation/n=10-2                            1.28µs ± 3%    1.27µs ± 1%     ~     (p=0.381 n=13+13)
PolygonSingleRingValidation/n=100-2                           19.2µs ± 2%    19.1µs ± 1%   -0.95%  (p=0.001 n=14+12)
PolygonSingleRingValidation/n=1000-2                           283µs ± 2%     283µs ± 1%     ~     (p=0.139 n=15+12)
PolygonSingleRingValidation/n=10000-2                         4.68ms ±23%    4.79ms ±20%     ~     (p=0.621 n=14+15)
PolygonMultipleRingsValidation/n=4-2                          3.65µs ± 1%    3.66µs ± 4%     ~     (p=0.659 n=12+13)
PolygonMultipleRingsValidation/n=36-2                         30.8µs ± 3%    30.7µs ± 4%     ~     (p=1.000 n=13+14)
PolygonMultipleRingsValidation/n=400-2                         407µs ± 3%     407µs ± 4%     ~     (p=0.982 n=14+14)
PolygonMultipleRingsValidation/n=4096-2                       5.18ms ± 1%    5.18ms ± 2%     ~     (p=0.720 n=13+14)
PolygonZigZagRingsValidation/n=10-2                           6.12µs ± 1%    6.13µs ± 3%     ~     (p=0.381 n=12+14)
PolygonZigZagRingsValidation/n=100-2                          74.1µs ± 2%    73.9µs ± 3%     ~     (p=0.164 n=14+14)
PolygonZigZagRingsValidation/n=1000-2                          989µs ± 5%     979µs ± 2%     ~     (p=0.981 n=14+13)
PolygonZigZagRingsValidation/n=10000-2                        16.3ms ±20%    16.2ms ±32%     ~     (p=0.838 n=15+15)
PolygonAnnulusValidation/n=10-2                               1.98µs ± 3%    1.98µs ± 5%     ~     (p=0.693 n=13+14)
PolygonAnnulusValidation/n=100-2                              19.2µs ± 4%    19.3µs ± 3%     ~     (p=0.650 n=13+15)
PolygonAnnulusValidation/n=1000-2                              347µs ± 2%     343µs ± 2%   -1.19%  (p=0.001 n=14+14)
PolygonAnnulusValidation/n=10000-2                            4.55ms ± 4%    4.55ms ± 4%     ~     (p=0.943 n=13+14)
MultipolygonValidation/n=1-2                                   218ns ± 4%     217ns ± 1%     ~     (p=0.092 n=13+14)
MultipolygonValidation/n=4-2                                   486ns ± 2%     486ns ± 2%     ~     (p=0.955 n=13+15)
MultipolygonValidation/n=16-2                                 2.16µs ± 4%    2.13µs ± 1%     ~     (p=0.091 n=14+13)
MultipolygonValidation/n=64-2                                 10.3µs ± 1%    10.3µs ± 2%     ~     (p=0.759 n=13+15)
MultipolygonValidation/n=256-2                                56.1µs ± 4%    56.1µs ± 4%     ~     (p=0.793 n=14+13)
MultipolygonValidation/n=1024-2                                342µs ± 2%     341µs ± 1%     ~     (p=0.899 n=14+12)
MultiPolygonTwoCircles/n=10-2                                 1.94µs ± 3%    1.92µs ± 2%     ~     (p=0.070 n=15+13)
MultiPolygonTwoCircles/n=100-2                                23.1µs ± 4%    22.9µs ± 1%     ~     (p=0.103 n=15+12)
MultiPolygonTwoCircles/n=1000-2                                298µs ± 8%     324µs ±28%     ~     (p=0.310 n=14+15)
MultiPolygonTwoCircles/n=10000-2                              4.88ms ±11%    4.82ms ±13%     ~     (p=0.567 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-2                      2.62µs ± 3%    2.60µs ± 1%     ~     (p=0.188 n=12+14)
MultiPolygonMultipleTouchingPoints/n=10-2                     20.9µs ± 2%    21.0µs ± 4%     ~     (p=0.847 n=13+15)
MultiPolygonMultipleTouchingPoints/n=100-2                     266µs ± 3%     265µs ± 1%     ~     (p=0.479 n=13+13)
MultiPolygonMultipleTouchingPoints/n=1000-2                   3.32ms ± 4%    3.28ms ± 1%   -1.22%  (p=0.048 n=13+14)
WKTParsing/point-2                                            1.13µs ± 3%    1.13µs ± 2%     ~     (p=0.659 n=13+13)
DistancePolygonToPolygonOrdering/n=100_swap=false-2           27.1µs ± 6%    26.8µs ± 6%     ~     (p=0.174 n=15+15)
DistancePolygonToPolygonOrdering/n=100_swap=true-2            26.6µs ± 8%    26.7µs ± 4%     ~     (p=0.252 n=12+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false-2           464µs ± 2%     464µs ± 2%     ~     (p=0.943 n=14+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true-2            464µs ± 2%     463µs ± 1%     ~     (p=0.977 n=12+12)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-2     3.25µs ± 2%    3.24µs ± 3%     ~     (p=0.343 n=13+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-2      3.26µs ± 5%    3.24µs ± 2%     ~     (p=0.676 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-2    35.1µs ± 4%    34.6µs ± 1%   -1.46%  (p=0.002 n=13+12)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-2     34.8µs ± 1%    34.9µs ± 4%     ~     (p=0.969 n=12+14)
MultiLineStringIsSimpleManyLineStrings/n=100-2                26.9µs ± 2%    26.7µs ± 1%   -0.55%  (p=0.014 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-2                331µs ± 3%     329µs ± 3%     ~     (p=0.194 n=14+14)
ForceCWandForceCCW/0-2                                        16.5ns ± 8%    16.5ns ± 5%     ~     (p=0.675 n=14+14)
ForceCWandForceCCW/0#01-2                                      103ns ± 3%     101ns ± 1%   -1.48%  (p=0.010 n=14+13)
ForceCWandForceCCW/1-2                                        16.9ns ± 6%    16.7ns ± 6%     ~     (p=0.462 n=14+12)
ForceCWandForceCCW/1#01-2                                      101ns ± 2%     102ns ± 2%     ~     (p=0.070 n=12+12)
ForceCWandForceCCW/2-2                                        31.1ns ± 4%    30.7ns ±10%     ~     (p=0.164 n=15+15)
ForceCWandForceCCW/2#01-2                                      155ns ± 3%     157ns ± 6%     ~     (p=0.167 n=14+14)
ForceCWandForceCCW/3-2                                        30.5ns ± 6%    30.5ns ± 4%     ~     (p=0.198 n=12+14)
ForceCWandForceCCW/3#01-2                                      156ns ± 3%     156ns ± 2%     ~     (p=0.571 n=10+12)
ForceCWandForceCCW/4-2                                        44.5ns ± 6%    43.9ns ± 5%     ~     (p=0.565 n=13+12)
ForceCWandForceCCW/4#01-2                                      241ns ± 3%     239ns ± 1%     ~     (p=0.110 n=10+10)
ForceCWandForceCCW/5-2                                        44.7ns ± 5%    44.5ns ± 6%     ~     (p=0.667 n=14+12)
ForceCWandForceCCW/5#01-2                                      241ns ± 3%     240ns ± 1%     ~     (p=0.936 n=12+10)
ForceCWandForceCCW/6-2                                        65.0ns ± 8%    63.7ns ± 6%     ~     (p=0.092 n=14+14)
ForceCWandForceCCW/6#01-2                                      392ns ± 6%     389ns ± 3%     ~     (p=0.960 n=13+13)
EnvelopeTransformXY-2                                          141ns ± 3%     142ns ± 4%     ~     (p=0.064 n=14+15)
WKBParse/0-2                                                   420ns ± 0%     420ns ± 1%     ~     (p=0.396 n=13+13)
WKBParse/1-2                                                   426ns ± 2%     425ns ± 1%     ~     (p=0.480 n=13+14)
WKBParse/2-2                                                   480ns ± 0%     491ns ± 1%   +2.25%  (p=0.000 n=12+14)
WKBParse/3-2                                                   484ns ± 5%     483ns ± 1%     ~     (p=0.990 n=13+14)
WKBParse/4-2                                                   558ns ± 1%     567ns ± 1%   +1.55%  (p=0.000 n=15+13)
WKBParse/5-2                                                   434ns ± 1%     424ns ± 1%   -2.31%  (p=0.000 n=13+15)
WKBParse/6-2                                                   487ns ± 2%     487ns ± 1%     ~     (p=0.923 n=14+15)
WKBParse/7-2                                                   486ns ± 0%     488ns ± 2%     ~     (p=0.368 n=13+15)
WKBParse/8-2                                                   577ns ± 1%     577ns ± 0%     ~     (p=0.461 n=13+12)
WKBParse/9-2                                                   313ns ± 2%     314ns ± 2%     ~     (p=0.310 n=13+15)
WKBParse/10-2                                                  315ns ± 5%     313ns ± 1%     ~     (p=1.000 n=13+14)
WKBParse/11-2                                                  313ns ± 1%     313ns ± 1%     ~     (p=0.708 n=13+15)
WKBParse/12-2                                                  308ns ± 1%     309ns ± 1%   +0.33%  (p=0.011 n=14+15)
WKBParse/13-2                                                  605ns ± 1%     595ns ± 1%   -1.64%  (p=0.000 n=13+14)
WKBParse/14-2                                                  704ns ± 1%     705ns ± 1%     ~     (p=0.877 n=13+14)
WKBParse/15-2                                                  713ns ± 1%     712ns ± 1%     ~     (p=0.381 n=13+14)
WKBParse/16-2                                                  836ns ± 1%     837ns ± 1%     ~     (p=0.253 n=13+14)
WKBParse/17-2                                                  703ns ± 1%     701ns ± 0%     ~     (p=0.883 n=14+14)
WKBParse/18-2                                                  907ns ± 0%     907ns ± 1%     ~     (p=0.883 n=12+13)
WKBParse/19-2                                                  908ns ± 1%     909ns ± 1%     ~     (p=0.553 n=13+13)
WKBParse/20-2                                                 1.09µs ± 0%    1.10µs ± 1%   +1.02%  (p=0.000 n=12+14)
WKBParse/21-2                                                  302ns ± 1%     301ns ± 1%   -0.49%  (p=0.018 n=14+14)
WKBParse/22-2                                                  303ns ± 1%     303ns ± 1%     ~     (p=0.305 n=14+15)
WKBParse/23-2                                                  304ns ± 1%     302ns ± 0%   -0.65%  (p=0.000 n=15+14)
WKBParse/24-2                                                  300ns ± 2%     298ns ± 1%     ~     (p=0.091 n=15+15)
WKBParse/25-2                                                 1.56µs ± 2%    1.56µs ± 1%     ~     (p=0.955 n=13+15)
WKBParse/26-2                                                 1.97µs ± 1%    1.97µs ± 1%     ~     (p=0.277 n=13+15)
WKBParse/27-2                                                 1.97µs ± 0%    1.98µs ± 1%     ~     (p=0.186 n=12+15)
WKBParse/28-2                                                 2.47µs ± 1%    2.47µs ± 1%     ~     (p=0.518 n=14+14)
WKBParse/29-2                                                  302ns ± 0%     301ns ± 1%   -0.29%  (p=0.023 n=13+14)
WKBParse/30-2                                                  304ns ± 2%     304ns ± 2%     ~     (p=0.174 n=14+14)
WKBParse/31-2                                                  303ns ± 1%     302ns ± 1%     ~     (p=0.075 n=14+15)
WKBParse/32-2                                                  299ns ± 1%     299ns ± 1%     ~     (p=0.246 n=14+15)
WKBParse/33-2                                                  569ns ± 1%     571ns ± 1%   +0.28%  (p=0.049 n=14+13)
WKBParse/34-2                                                  660ns ± 0%     649ns ± 1%   -1.72%  (p=0.000 n=14+15)
WKBParse/35-2                                                  659ns ± 0%     659ns ± 1%     ~     (p=0.733 n=12+14)
WKBParse/36-2                                                  722ns ± 2%     711ns ± 1%   -1.47%  (p=0.000 n=13+14)
WKBParse/37-2                                                  807ns ± 1%     814ns ± 0%   +0.83%  (p=0.000 n=13+12)
WKBParse/38-2                                                  965ns ± 1%     962ns ± 3%     ~     (p=0.105 n=13+15)
WKBParse/39-2                                                  965ns ± 1%     966ns ± 1%     ~     (p=0.910 n=13+13)
WKBParse/40-2                                                 1.07µs ± 2%    1.08µs ± 1%   +0.86%  (p=0.000 n=14+14)
WKBParse/41-2                                                  303ns ± 1%     302ns ± 0%   -0.38%  (p=0.000 n=13+14)
WKBParse/42-2                                                  306ns ± 1%     305ns ± 1%     ~     (p=0.489 n=14+14)
WKBParse/43-2                                                  305ns ± 1%     306ns ± 1%     ~     (p=0.802 n=13+14)
WKBParse/44-2                                                  300ns ± 1%     302ns ± 1%   +0.40%  (p=0.009 n=13+15)
WKBParse/45-2                                                  888ns ± 1%     889ns ± 3%     ~     (p=0.777 n=14+14)
WKBParse/46-2                                                 1.09µs ± 1%    1.08µs ± 1%   -0.61%  (p=0.023 n=15+13)
WKBParse/47-2                                                 1.08µs ± 2%    1.08µs ± 2%     ~     (p=0.839 n=14+13)
WKBParse/48-2                                                 1.23µs ± 2%    1.24µs ± 1%   +0.71%  (p=0.011 n=14+12)
WKBParse/49-2                                                 1.28µs ± 1%    1.27µs ± 2%   -0.68%  (p=0.002 n=15+13)
WKBParse/50-2                                                 1.60µs ± 1%    1.60µs ± 1%     ~     (p=0.711 n=14+13)
WKBParse/51-2                                                 1.59µs ± 1%    1.61µs ± 3%   +1.18%  (p=0.000 n=12+13)
WKBParse/52-2                                                 1.87µs ± 1%    1.88µs ± 1%   +0.83%  (p=0.000 n=14+12)
WKBParse/53-2                                                  303ns ± 1%     302ns ± 1%   -0.56%  (p=0.000 n=14+13)
WKBParse/54-2                                                  305ns ± 1%     305ns ± 1%     ~     (p=0.103 n=14+12)
WKBParse/55-2                                                  305ns ± 1%     305ns ± 4%     ~     (p=0.181 n=13+13)
WKBParse/56-2                                                  300ns ± 1%     299ns ± 1%     ~     (p=0.135 n=13+13)
WKBParse/57-2                                                 1.93µs ± 1%    1.92µs ± 2%   -0.41%  (p=0.005 n=12+13)
WKBParse/58-2                                                 2.38µs ± 1%    2.37µs ± 1%   -0.38%  (p=0.047 n=15+13)
WKBParse/59-2                                                 2.37µs ± 1%    2.36µs ± 1%   -0.62%  (p=0.000 n=13+12)
WKBParse/60-2                                                 2.86µs ± 1%    2.86µs ± 2%     ~     (p=0.990 n=13+14)
WKBParse/61-2                                                  302ns ± 1%     303ns ± 1%     ~     (p=0.641 n=14+13)
WKBParse/62-2                                                  305ns ± 2%     304ns ± 1%     ~     (p=0.838 n=15+14)
WKBParse/63-2                                                  304ns ± 1%     303ns ± 1%     ~     (p=0.334 n=14+14)
WKBParse/64-2                                                  298ns ± 1%     299ns ± 1%     ~     (p=0.077 n=15+13)
WKBParse/65-2                                                  871ns ± 0%     859ns ± 1%   -1.41%  (p=0.000 n=13+14)
WKBParse/66-2                                                 1.03µs ± 1%    1.02µs ± 1%   -0.46%  (p=0.002 n=14+13)
WKBParse/67-2                                                 1.01µs ± 1%    1.02µs ± 1%   +0.76%  (p=0.000 n=14+13)
WKBParse/68-2                                                 1.14µs ± 1%    1.13µs ± 1%   -0.37%  (p=0.007 n=13+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:arm64
IntersectionWithoutValidation/n=10-2                          13.4µs ± 2%    13.4µs ± 2%     ~     (p=0.475 n=14+14)
IntersectionWithoutValidation/n=100-2                         29.1µs ± 2%    29.0µs ± 1%     ~     (p=0.652 n=15+14)
IntersectionWithoutValidation/n=1000-2                         149µs ± 3%     149µs ± 3%     ~     (p=0.801 n=13+13)
IntersectionWithoutValidation/n=10000-2                       1.38ms ± 5%    1.41ms ± 4%   +2.37%  (p=0.025 n=14+13)
NoOp/n=10-2                                                   2.17µs ± 2%    2.16µs ± 2%     ~     (p=0.290 n=14+14)
NoOp/n=100-2                                                  5.70µs ± 1%    5.73µs ± 4%     ~     (p=0.395 n=14+14)
NoOp/n=1000-2                                                 39.8µs ± 2%    40.3µs ± 5%     ~     (p=0.367 n=15+15)
…